Move AgentContext to Ignition.Core.Mvc namespace
